Rename more files/components
