+++ src/search/jobsIndexNames.js
@@ -1,0 +1,2 @@
+const jobsIndexNames = ['name', 'id', 'harvestableId', 'message', 'all'];
+export default jobsIndexNames;
