--- src/search/oldJobsIndexNames.js
@@ -1,2 +1,0 @@
-const oldJobsIndexNames = ['name', 'id', 'harvestableId', 'message', 'all'];
-export default oldJobsIndexNames;
